Brings in all work from the fb-terrm framebuffer terminal session:

- kernel_entry.S: fix SysV ABI stack alignment (sub rax, 8 before jmp)
- kernel_main.c: remove redundant kmalloc stack switch
- framebuffer.h/c: FB_ANSI_PALETTE visibility("hidden") — fixes GOT/PE
  crash in apply_sgr during EXEC-DOE with framebuffer active
- tools/fbtest.c: host render harness + PNG output + 8 pixel assertions
- scripts/qemu_screenshot.sh: QEMU HMP screendump → PNG
- Makefile.starkernel: fbtest + screendump wired into amd64 qemu target
- docs/lithosananke/kernel-args/DESIGN.md: KernelArgs design (JVM-style
  UEFI command-line overrides for stack, heap, --doe, --log-level)
- docs/NEXT_SESSION.md: task queue for next session
- CLAUDE.md: one-process-at-a-time build rule

EXEC-DOE now runs to full completion (160,673 rows) without GP fault.
amd64 acceptance confirmed clean.

+++ src/starkernel/doe_log.c
@@ -26,8 +26,8 @@
  *   7.  hot_word_count         uint64  — words with heat >= threshold
  *   8.  avg_word_heat_q48      uint64  — mean execution heat as raw Q48.16 integer
  *   9.  window_width           uint32  — rolling window effective size
- *   10. actual_window_size     uint32  — ring buffer fill level
- *   11. predicted_label_hits   uint32  — context prediction hits (TODO: 0)
+ *   10. actual_window_size     uint32  — true analysis width: min(total_executions, effective_window_size)
+ *   11. predicted_label_hits   uint32  — ANOVA early-exit confirmations per tick (L8 validation signal)
  *   12. jitter_bits            uint64  — estimated_jitter_ns IEEE 754 raw bits (union-punned)
  *   13. apic_ticks             uint64  — APIC timer monotonic tick count
  *   14. time_trust_q48         uint64  — TIME-TRUST in Q48.16 format
@@ -83,25 +83,27 @@ void doe_log_tick_row(VM *vm, const HeartbeatTickSnapshot *snap)
     jitter_bits.d = snap->estimated_jitter_ns;
     uint64_t jitter_raw = jitter_bits.u;
 
-    /* Format the 15-column CSV row into a local buffer */
+    /* Format the 15-column CSV row into a local buffer.
+     * Use %llu (unsigned long long) for all uint64_t fields — %lu is unreliable
+     * for values > 2^32 on aarch64 due to mixed-width varargs ABI behaviour. */
     char buf[DOE_BUF_SIZE];
     snprintf(buf, sizeof(buf),
-             "%u,%lu,%lu,%u,%u,%u,%lu,%lu,%u,%u,%u,%lu,%lu,%lu,%lu",
+             "%u,%llu,%llu,%u,%u,%u,%llu,%llu,%u,%u,%u,%llu,%llu,%llu,%llu",
              snap->tick_number,
-             snap->elapsed_ns,
-             snap->tick_interval_ns,
+             (unsigned long long)snap->elapsed_ns,
+             (unsigned long long)snap->tick_interval_ns,
              snap->cache_hits_delta,
              snap->bucket_hits_delta,
              snap->word_executions_delta,
-             snap->hot_word_count,
-             avg_heat_q48,
+             (unsigned long long)snap->hot_word_count,
+             (unsigned long long)avg_heat_q48,
              snap->window_width,
              snap->actual_window_size,
              snap->predicted_label_hits,
-             jitter_raw,
-             apic_ticks,
-             time_trust_q48,
-             variance_q48);
+             (unsigned long long)jitter_raw,
+             (unsigned long long)apic_ticks,
+             (unsigned long long)time_trust_q48,
+             (unsigned long long)variance_q48);
 
     console_puts(DOE_PREFIX);
     console_puts(buf);
